Made use of size_t for better 64-bit support.

+++ include/libbndl/bundle.hpp
@@ -19,8 +19,8 @@ namespace libbndl
 
 		struct Entry
 		{
-			uint32_t	Size;
-			uint32_t	Offset;
+			size_t	Size;
+			size_t	Offset;
 		};
 	public:
 		bool Load(const std::string& name);
@@ -29,14 +29,14 @@ namespace libbndl
 		//get entry info
 		Entry GetInfo(const std::string& entry);
 		//get entry binary
-		uint8_t* GetBinary(const std::string& entry, uint32_t& size);
+		uint8_t* GetBinary(const std::string& entry, size_t& size);
 		//get entry as string
 		std::string GetText(const std::string& entry);
 		//get entry stream
 		std::istream GetStream(const std::string& entry);
 		//add entry 
-		void AddEntry(const std::string& entry, const std::string& text,bool overwrite = true);
-		void AddEntry(const std::string& entry, const uint8_t* data, uint32_t size, bool overwrite = true);
+		void AddEntry(const std::string& entry, const std::string& text, bool overwrite = true);
+		void AddEntry(const std::string& entry, const uint8_t* data, size_t size, bool overwrite = true);
 
 		std::vector<std::string> ListEntries();
 	private:
